[lightify] fix start up of connections to gateways
Sometimes when connecting to a gateway every requests seems to result in
an 0x16 status response until a LIST_PAIRED_DEVICES is sent. It seems a
LIST_PAIRED_DEVICES is used as some sort of resync procedure and an 0x16
status is a request, a pretty insistent request, from the gateway to
resync.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)