[lightify] Synchronize the various group channels
A group has more channels than is strictly necessary because creating a
group of, say, power switches and then having to link switch Items to the
color channel makes no sense to users. But we do need to keep the channels
in sync because they reflect the same command path to devices so, for
instance, an "ON" is _just_ an "ON" and turns on both lights and power
switch members regardless of which channel it was sent to. If we don't
synchronize the channels it gives the impression that they act
independently on just those things they are "obviously" relevant to and
that just isn't the case.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)